attached_assets/Pasted-Uncaught-runtime-errors-ERROR-Element-type-is-invalid-expected-a-string-for-built-in-compone-1741468381947.txt

Uncaught runtime errors:
×
ERROR
Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `ParticipantsViewer`.
Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `ParticipantsViewer`.
    at createFiberFromTypeAndProps (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:93247:21)
    at createFiberFromElement (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:93268:19)
    at reconcileSingleElement (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80862:27)
    at reconcileChildFibers (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80912:39)
    at reconcileChildren (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:85342:32)
    at mountIndeterminateComponent (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:86186:9)
    at beginWork (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:87421:20)
    at HTMLUnknownElement.callCallback (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:72403:18)
    at Object.invokeGuardedCallbackDev (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:72447:20)
    at invokeGuardedCallback (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:72504:35)
ERROR
Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `ParticipantsViewer`.
Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `ParticipantsViewer`.
    at createFiberFromTypeAndProps (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:93247:21)
    at createFiberFromElement (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:93268:19)
    at reconcileSingleElement (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80862:27)
    at reconcileChildFibers (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80912:39)
    at reconcileChildren (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:85342:32)
    at mountIndeterminateComponent (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:86186:9)
    at beginWork (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:87421:20)
    at HTMLUnknownElement.callCallback (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:72403:18)
    at Object.invokeGuardedCallbackDev (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:72447:20)
    at invokeGuardedCallback (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:72504:35)
ERROR
Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `ParticipantsViewer`.
Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `ParticipantsViewer`.
    at createFiberFromTypeAndProps (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:93247:21)
    at createFiberFromElement (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:93268:19)
    at reconcileSingleElement (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80862:27)
    at reconcileChildFibers (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80912:39)
    at reconcileChildren (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:85342:32)
    at mountIndeterminateComponent (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:86186:9)
    at beginWork (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:87421:20)
    at beginWork$1 (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:92380:18)
    at performUnitOfWork (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:91650:16)
    at workLoopSync (https://6932e9bc-76a6-4252-b635-94603e2442ea-00-3k2njy9325e91.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:91573:9)
